Move out broadcast logic and perform PR cleanup
